♻️ Extract the implementations about core processes of GitHub label operations into new module.
